refactor(kernel): use isType utilities for fn & string
