+++ packages/kernel/src/eventaggregator.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
 import { DI } from './di.js';
 import { Constructable, IDisposable } from './interfaces.js';
+import { isString } from './utilities.js';
 
 /**
  * Represents a handler for an EventAggregator event.
@@ -56,7 +57,7 @@ export class EventAggregator {
       throw new Error(`Invalid channel name or instance: ${channelOrInstance}.`);
     }
 
-    if (typeof channelOrInstance === 'string') {
+    if (isString(channelOrInstance)) {
       let subscribers = this.eventLookup[channelOrInstance];
       if (subscribers !== void 0) {
         subscribers = subscribers.slice();
@@ -107,7 +108,7 @@ export class EventAggregator {
     let handler: unknown;
     let subscribers: unknown[];
 
-    if (typeof channelOrType === 'string') {
+    if (isString(channelOrType)) {
       if (this.eventLookup[channelOrType] === void 0) {
         this.eventLookup[channelOrType] = [];
       }
