--- packages/kernel/src/shared.ts
@@ -1,8 +1,0 @@
-import { Metadata } from '@aurelia/metadata';
-
-/** @internal */
-export const getOwnMetadata = Metadata.getOwn;
-/** @internal */
-export const hasOwnMetadata = Metadata.hasOwn;
-/** @internal */
-export const defineMetadata = Metadata.define;
